chore: move faq into component
